manatee: limit dentry/inode cache
By default, the dentry cache allocates a hash table with num_pages/2
buckets, rounded up to a power of 2. On an 8GiB system, this works out
to 1048576 buckets, which at 8 bytes per bucket consumes 8MiB of memory.
The inode cache is similar, except the hash table is half as large.

Specify a smaller limit on manatee, since not much will run in the
hypervisor. The chosen values are 1/8 the default size for a 8GB system.
These values were chosen somewhat arbitrarily - they are 15-20x the
number of dentry/inode objects on my current test device. With these
values, hypervisor memory consumption is reduced by ~10MiB on an 8GB
device and ~22Mib on a 16GB device.

BUG=b:218423936
TEST=CQ

Change-Id: Iebbaa9be0efb389fad04df0735c9d145d6857dca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3445709
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>
Tested-by: David Stevens <stevensd@chromium.org>