trogdor: Fix misplaced zephyr_firmware_common in model.yaml
This wasn't doing anything as it was misplaced (specifying "firmware"
at the device level is not valid, it must be under "config").

Note: this does not affect any shipping trogdor images.  It's only
enabled for the "trogdor-zephyr" board based on the zephyr_ec USE
flag.

BUG=b:218868887
TEST=emerge-trogdor-zephyr -j chromeos-config
     observe zephyr-ec is set for lazor & limozeen

Change-Id: I612aee7ff89e7085e05551e81901a3cfeaa19e05
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3457482
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Bob Moragues <moragues@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>