overlay/profiles/base: Disable on-device handwriting on 2gb boards
This CL ensures the libhandwriting is disabled for all 2gb boards
as they don't support it. This is a prerequisite to an upcoming CL
where we enable libhandwriting on all the boards where the ml
service is enabled, so that 2gb boards will stay excluded.
The "-ondevice_handwriting" was added to all the overlays from
crrev/c/3239707 that are still in the repo; that is, excluding:
auron, buddy, gandof and lulu.

BUG=b:225063137
TEST=No affected tests because the flag was disabled by default.

Change-Id: Iecb1411854f9aec919d02fa4995c9860e78f532d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3541445
Reviewed-by: Honglin Yu <honglinyu@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Dmitry Grebenyuk <dgrebenyuk@google.com>
Commit-Queue: Dmitry Grebenyuk <dgrebenyuk@google.com>