scarlet: Add a device config with no SKU ID
Early scarlet RO firmware did not support exposing a SKU ID via
device-tree.

Historically, this was hacked as a one-off case into mosys [1], but
was never integrated into the rest of cros_config.  When the mosys
implementation of cros_config probing was deleted, so was this one-off
hack.

The mosys hack should have never landed.  Encoding a config which does
not require a SKU ID is as simple as just defining one in model.yaml.
This CL does that.

[1]: https://crrev.com/c/1284810

BUG=chromium:1300690
TEST=emerge-scarlet chromeos-config
     -> observe dru with no SKU ID at the end of the config list

Change-Id: I0d151fbea63f175870f4be08f421f845e5c3e256
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3544431
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: YH Lin <yueherngl@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>