guado: Disable V4L2.
In Chrome, we currently assume that only one of the v4l2_codec or vaapi
USE flags is set. This assumption does not apply to guado because guado
can use both APIs. When both USE flags are set, we get build errors. We
didn't care about this because guado had reached AUE. However, we should
still be able to build guado-labstation images. Video codec acceleration
is not important for these images, so this CL simply disables one of the
APIs on guado.

Bug: b:226995633
Test: cros tryjob -g 3558629 guado_labstation-release-tryjob
Change-Id: If1532e68041f839cbd586e59f6873f05d2c5ce69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3558629
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>