RK3399: Remove PlaybackMixerElem of speakers
When the mixer control "Speakers Switch" is 0, the audio playback will
not work even if we set the value back to 1 during playback.

The only case to trigger this issue is to mute speaker before playing
audio. We can solve it by removing PlaybackMixerElem of speaker from UCM
config. Therefore, the "Speakers Switch" will not be set to 0 when the
audio is muted.

This change will not affect the mute function since CRAS has already
handled it in the user space. In addition, there is no volume control of
speakers so that this change will not affect the volume.

BUG=b:221377113
TEST=1. Play audio from the internal speaker.
2. Mute the internal speaker.
3. Switch to the 3.5mm headphone and then switch back.
4. Unmute the internal speaker and make sure the audio sounds normal.

Change-Id: I2ef9e90f38ca012c9a2a1a124d44e4ce2ed60398
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3660713
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>