trogdor: Use TEO cpuidle governor by default
The kernel doesn't have a Kconfig option to set the default cpuidle
governor. Instead, we pass the argument on the kernel commandline. On
chromeos-5.4 we disabled the menu governor in the kernel config to get
TEO to be used. On chromeos-5.10 menu was enabled again, leading to TEO
being dropped in favor of menu because of cpuidle governor rating
differences. We'd like to keep both governors enabled in case the menu
governor proves to be more favorable on trogdor. This also makes sure
TEO is used even when the menu governor is compiled into the image, e.g.
upstream kernel configs.

For now, use TEO because the governor improves boot times and matches
the cpuidle governor we've been using on trogdor with chromeos-5.4.

BUG=b:239626992
TEST=grep teo /sys/devices/system/cpu/cpuidle/current_governor

Change-Id: I57a76281f27b9565cf12651cc9a28b0dfec3d767
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3785187
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>