overlays: Enable kernel AFDO on kukui
It was disabled for the sake of A/B test to measure AFDO efficiency
on Arm kernels. The goal was to compare performance stats
collected from the field on jacuzzi (where AFDO is enabled) and kukui.
The experiment is done and we can re-enable AFDO.

BUG=b:218702582
TEST=cq

Change-Id: I5748ea0389e40feaf13489420f8834ff1d4ddf69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3795052
Commit-Queue: Denis Nikitin <denik@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Tested-by: Denis Nikitin <denik@chromium.org>