baseboard-rex: Enable kernel messages on the UART
For the foreseeable future, we expect that the display / graphics on
MTL may continue to go through a cyle of breakages until a large
number of patches we are forced to carry, are actually committed to
the kernel. Lets enable the kernel messages on UART so that everyone
can use it to determine if all is well, or debug if needed.

BUG=b:242224323
TEST=build

Change-Id: Ib30cf882d75a03e12d10cecab47724321a648cac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3836496
Auto-Submit: Rajat Jain <rajatja@google.com>
Commit-Queue: YH Lin <yueherngl@chromium.org>
Commit-Queue: Rajat Jain <rajatja@google.com>
Reviewed-by: Rajat Jain <rajatja@google.com>
Tested-by: Rajat Jain <rajatja@google.com>
Tested-by: YH Lin <yueherngl@chromium.org>