corsola: Set default governor to schedutil
This equivalently reverts CL:3657331.

The schedutil governor aims at better integration with Linux kernel
scheduler, replacing the existing ondemand governor and is preferred
for big.LITTLE core systems.

The schedutil governor is reliant on the "boost-urgent" cros-config
for better performance on high priority tasks, and we already set that
for Corsola.

We revisited the benchmarks recently (in M113), and it looks like we're
okay to make the switch without too much performance impact.

BUG=b:245804557,b:233325718
TEST=cpufreq_config
     cat /sys/devices/system/cpu/cpu*/cpufreq/scaling_governor

Cq-Depend: chrome-internal:4983506
Change-Id: Ie3341109300f6cef8bb37403dd2bf8ff5339cab2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3905765
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>
Reviewed-by: Yu-Che Cheng <giver@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>