overlay-simple-fake-overlay: Add a new low-op board.
Add a new simple, standalone overlay. It is meant to eventually
have a depgraph with representative dependencies, but where every
package build is a noop or very low-op, so build_packages and
build_image can be completed in a handful of minutes rather than
an hour+.

The overlay inherits from amd64-generic to ensure toolchain prebuilts,
but otherwise should utilize as little as possible from the other
overlays.

BUG=b:241985342
TEST=build_packages && build_image

Change-Id: I702c97a689e692e9977f985018b4fd45f780e4aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4082441
Commit-Queue: Alex Klein <saklein@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Alex Klein <saklein@chromium.org>