chipset-mt8188g: Use 5.15 kernel temporarily
Note that 5.15 kernel is not the final kernel version.
We're targeting 6.0 kernel but the repo is not ready, and we use the
existing 5.15 kernel repo as a temporary place to put our kernel ToT
squash.

This will be changed later when the correct repo is created.

BUG=b:239660108,b:245739330
TEST=CQ

Change-Id: Icbe651fc2df2f634ddabd3ba0bc2661cb4c722c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3932511
Tested-by: Fei Shao <fshao@chromium.org>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>