baseboard-brya: move opengles metapackages from baseboard-brya-private
Pulls in the mesa-iris package, which is required on brya-derived
boards, as opposed to the the usual mesa package.

These ebuilds are currently in the baseboard-brya-private overlay, but
they are needed here to prevent file collisions between mesa and
mesa-iris when using the public manifest.

BUG=b:233387633
TEST=build_packages --board=brya  # using public repo manifest

Change-Id: I90911bea3edc7ce3af4aab94729927e82512468e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3935991
Reviewed-by: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-by: YH Lin <yueherngl@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: Ilja Friedel <ihf@chromium.org>