overlays: Don't hardcode mmc/mmcblk indices
These are kernel implementation details, and subject to change for a
variety of reasons. We rely on hardware addresses for the interesting
parts of these paths.

Note that these are shell expansions that end up partly used in
'get_fixed_dst_drive' (and therefore chromeos-install, etc.), but we
only care that the last element of the path expands correctly. Here be
dragons.

Also note that due to the aforementioned dragons, this "seemingly
correct" change breaks some boards which had benignly-incorrect
rootdev_paths. See CL:4013883.

BUG=b:258046147
TEST=CQ
TEST=`chromeos-install` (without --dst) on a few

Cq-Depend: chromium:4013883
Change-Id: I23d331c70f2958acd10b17c0971c1ca9b9f90b4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4010499
Reviewed-by: George Engelbrecht <engeg@google.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>