reven: Add initial public overlay
This board is mostly empty for now, just the minimal set of necessary
files to build a new board. This board will be used as the new parent
board for `reven-private`, and we can shift parts of the private overlay
that are OK to make public incrementally.

The files added here are mostly identical to the ones currently in
overlay-reven-private, except for copyright header updates and some
reven-private -> reven changes. The exception is `make.defaults`, which
has a couple USE flags needed to make build_packages work, but is
otherwise empty. Note that the board's appid is the same as
reven-private, which as far as I can tell is the usual pattern.

This is not yet a usable board; in a VM it only boots as far as the
splash before resetting.

BUG=b:246558823
TEST=Create source checkout with the public (not internal) manifest
TEST=setup_board --board=reven --default
TEST=cros_workon start chromeos-bsp-reven
TEST=build_packages && build_image test

Change-Id: I39121f8a090c0908126e889db31aa23205602a96
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4015019
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Reviewed-by: Chris Gerber <gerb@google.com>