board-overlays: Fix rootdev_base for several ARM systems
Similar to
https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4013883,
which had to get reverted because of its parent CL. Let's do this first
this time.

The rootdev_base is wrong for many systems:

* /sys/devices/soc (for platform devices) is at
  /sys/devices/platform/soc on recent kernels.

* on upstream (5.11) kernels, RK3399/Gru systems changed their
  SDHCI/eMMC node from "<address>.sdhci" to "<address>.mmc":

    9a9f64278407 arm64: dts: rockchip: rename sdhci nodename to mmc on rk3399

  Use a wildcard to handle that.

* The TODO address on Asurada was wrong (not 11230000; it should be
  11f60000) -- see
  src/third_party/kernel/v5.4/arch/arm64/boot/dts/mediatek/mt8192.dtsi

Note that this mostly doesn't matter per b/258046147#comment8, because
in practice we only care that the last element of the path resolves to
something in /dev/ after shell expansion. But these incorrect paths
leave room for error on future refactoring.

While correcting these, add wildcards in the appropriate places per
b/258046147.

BUG=b:150746591, b:109850061, b:258046147
TEST=platform.PartitionSizes on various systems

Change-Id: I677a600c0ded88ecfb17f7b377d2d4362e4eda46
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4024348
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>