baseboard-coral: Don't hardcode mmc/mmcblk indices
This is the main fix extracted from:

https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4010499

because that CL had to be reverted for other breakages (it changed a lot
of other overlays).

This particular change is critical, because Coral devices have two SDHCI
hosts, and with asynchronous probe (new on M-109), those two hosts may
get different numbering. We need to be resilient.

BUG=b:258046147
TEST=platform.PartitionSizes

Change-Id: Ie11f819403810068468a586ff66d9882c2407cec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4024347
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Tested-by: Brian Norris <briannorris@chromium.org>