nissa: Remove chromeos-ish ebuild
For nissa, the ISH main firmware will be included in the CSE region of
the firmware image instead of loading it from rootfs. So it doesn't need
to be installed in /lib/firmware, it just needs to be installed
somewhere in /firmware so that it will be built and included in the
firmware archive. It is then manually stitched into the CSE binary
using scripts in coreboot-private-files-chipset-adln.

The chromeos-ec ebuild already builds and installs it as
/firmware/adl_ish_lite/ec.bin, so the chromeos-ish ebuild doesn't need
to build it too.

Remove the chromeos-ish ebuild and disable the cros_ish USE flag. Move
the ish.conf installation to chromeos-bsp-nissa.

BUG=b:234776154
TEST=`emerge-nissa chromeos-bsp-nissa`, check /etc/modprobe.d/ish.conf
is installed.
`emerge-nissa chromeos-ec`, check ISH firmware is still built and
installed as /firmware/adl_ish_lite/ec.bin

Change-Id: Idaa074d730330f1d5776d41b73a0f6c0302c4973
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4044471
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Tested-by: Reka Norman <rekanorman@chromium.org>