snappy: Remove unnecessary usage of "mosys platform vendor" in oemdata.sh
The only three models supported by the snappy board (alan, bigdaddy,
and snappy) are all HP OEM, so the check if the system vendor is HP is
a tautology.

For context, the "mosys platform vendor" command is deprecated, and
will be removed in the near future.

BUG=b:187790074
TEST=CQ

Change-Id: I563314daa3cc8549df88d5bf8ecafdc106d5b950
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4067441
Reviewed-by: YH Lin <yueherngl@chromium.org>
Commit-Queue: YH Lin <yueherngl@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>