Change behavior on boards boards to use the intrinsic gain
This CL removes the gains applied in APM for Redrix, Gimble, Anahera and Vell. This is now possible, and needed, since the move of the intrinsic gain to be applied in front of APM in CRAS is completed, and the gain applied via apm.ini is no longer needed or wanted.

BUG=b:191852990
TEST=Manually tested on a Redrix device

Change-Id: I3daa60ced4761c044a39a4072e39774412ca528f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4100688
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Tested-by: Li-Yu Yu <aaronyu@google.com>
Tested-by: Per Åhgren <peah@chromium.org>
Reviewed-by: Per Åhgren <peah@chromium.org>
Commit-Queue: Per Åhgren <peah@chromium.org>
Auto-Submit: Per Åhgren <peah@chromium.org>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>