trogdor: Fix verify_fsg to handle kernel 5.15
When we uprevved to kernel 5.15, the sysfs path for the efuses
changed. It was:
  /sys/bus/nvmem/devices/qfprom0/nvmem
It's now:
  /sys/bus/nvmem/devices/qfprom1/nvmem

Let's change the script to use a wildcard. To do this, we use
`realpath` to resolve the wildcard since otherwise the just assigning
the variable includes the asterisks.

BUG=b:265491525
TEST=Put this on a fused device and confirm fsg corruption is detected

Change-Id: I495c7aedbe87cbe52303c845d5aa9f55a1a86b62
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4179670
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>