trogdor: Shorten the EC's hibernation delay from 1 hour to 10 seconds
We are seeing cases where lazor devices stop detecting tablet mode
properly. Typically, users are reporting that they are stuck in tablet
mode and their keyboard/trackpad don't work at all. An EC reset
(Escape-Refresh) is needed to resolve the problem.

It would be good to understand how this is happening, but so far all
attempts to spot the bug have not been fruitful and we have no way to
reproduce the bug. We've added logging to get some clues, but that
logging won't roll out until the next firmware spin.

As a stopgap, let's significantly decrease the timeout before we enter
hibernation. After hibernation exit, the sensors will have been power
cycled and the EC's state will start again nearly from scratch. Any
tablet mode issues will be resolved.

After this change, users should be able to resolve this problem by
"turning it off and on again". Specifically, a user could fix
themselves by:
1. Making sure they're not plugged into power.
2. Powering off.
3. Counting slowly to 10.
4. Powering back on.

That seems like something that a user might think to try themselves
and hopefully will at least get users back to a working laptop.

NOTE: one might ask why we originally had the hibernation delay as 1
hour when 10 seconds would do fine. There's not any really strong
reason here. The difference between G3 and hibernation is:
* The device will take slightly longer to bootup.
* Some wakeup sources are different. For instance, in hibernate
  _closing_ the lid will wake you up (though we'll just turn back off
  again if we boot up with the lid closed). Also, in hibernate
  plugging in or pulling out power (with the lid open) will turn the
  device on.
* Hibernate takes up a bit less power, though really this matters when
  a device is going to be off for an extended period of time.

In general, though, if the end user experience is fine when the device
has been off for an hour, it should also be fine for when the device
is off for 10 seconds.

This current implementation is via a udev rule and a direct call to
"ectool". This seems like the most expedient way to get this
landed. "ectool" is already called in a few places in shipping images
(including to populate some stuff in chrome://system). Trying to
change the default in the EC requires EC spins. Trying to come up with
a way to do this in the kernel seems quite daunting.

NOTE: at the moment, most analysis of this problem has been done on
lazor. However, there's no reason to believe that all trogdor boards
couldn't be affected in the same way. We'll apply the rule
universally.

BUG=b:267680317
TEST=See hibernate get set to 10 seconds after boot

Change-Id: I5e0a65ca36e272eeab0740a1f590a8639c2c487d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4300920
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Commit-Queue: Douglas Anderson <dianders@chromium.org>