chipset-adl: Fix zavala encoder fps expectations
The CL/4241237 has increased all the brya boards FPS expectations for
some of the decoder and the encoder tests.

As the brya.zavala was redefining only the decoder values, the encoder
expectations were increased for this board as well. Unfortunately, the
zavala needs to be treated separately in terms of the performance.

The xml specifying the zavala decoder performance expectations was
updated with the general perf numbers prior to CL/4241237 for the
failing tests.

BUG=b:271676475
TEST=android.video.cts.VideoEncoderDecoderTest#testAvcOther0Perf0320x0240
TEST=android.video.cts.VideoEncoderDecoderTest#testAvcOther0Perf0720x0480

Change-Id: I9a903550974d15e59ffb050ea454037f6649468f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4306749
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Michał Krawczyk <mikrawczyk@google.com>
Reviewed-by: YH Lin <yueherngl@chromium.org>
Commit-Queue: Marcin Wojtas <mwojtas@google.com>
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>