trogdor/herobrine: Move esim USE flag to public baseboard
Right now we have the "esim" USE flag scattered between a bunch of
private overlays. Really, we want this in the baseboard to avoid
copying it. It's also nice to make it in the public overlays. There's
nothing secret about this USE flag, and even if the "hermes" ebuild is
in the private overlays it still doesn't hurt to set the USE flag in
the public overlays.

BUG=b:268332853
TEST=CQ

Change-Id: Id212d0f235995fbdfa6963b297e12e5392707b69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4330481
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>