overlay-zork: Add "clearcpuid=xsaves"
We (The AMD Chromeos team) believe that AMD's Erratum 1387 applies to
Picasso, Dali, and Pollock processors. The revision guide for these
chips has not been updated to include this information, but the RV1
processor (Zen 1) revision guide says that it's affected, and the issue
is present even up to the Zen3 processors like Cezanne, though it's
fixed there with a microcode update.

Because Linux uses the XSAVES command (Windows apparently does not),
this could possibly affect ChromeOS, though it's not clear to me how or
where the xsaves instruction is used. To work around any possible issue,
if we clear the xsaves feature flag from the CPUID, Linux should use
alternate methods (XSAVEOPT or XSAVEC) to save the XMM registers.

This is done on the linux kernel command line by adding
"clearcpuid=xsaves".

Note that this flag will clear the xsaves CPUID bit for anything that
gets the bit through the kernel, but it does not affect the hardware
CPUID itself, so reading the flags with iotools for example will not
show the bit as being cleared. /proc/cpuinfo should show the flag as
being cleared however.

BUG=b:271817196
TEST=CQ
TEST=Boot zork device, verify parameter is enabled.

Change-Id: Ibb59b495282a1381deaeea2643f0331033b17a9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4337174
Reviewed-by: Mark Hasemeyer <markhas@google.com>
Tested-by: Tim Van Patten <timvp@google.com>
Commit-Queue: Tim Van Patten <timvp@google.com>
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>