trogdor: Set scheduler-tune/boost-top-app to 60
The "boost-top-app" attribute in "scheduler-tune" is used for ARCVM to
give a boost to the "top app" in Android.

This is needed after CL:4353352 in order to get trogdor back to a
boost of 60. Prior to that CL "boost-urgent" was actually used for
ARCVM's "top app" boost. After that CL we need to specify the Android
boost explicitly.

BUG=b:217825939, b:243336790, b:248028313
TEST=cros_config /scheduler-tune boost-top-app

Change-Id: Id270fb369dbdfb38b75c27cf9893569ab04b14fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4354693
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Rob Clark <robdclark@chromium.org>
Reviewed-by: Masami Hiramatsu <mhiramat@google.com>