brya: cras: disable usage of hw echo ref
It is a known issue that the latency of the echo ref on the Max98373 is
higher than the latency of the DMIC audio. This results in a very poor
echo cancellation environment. Even though the amp modifies the signal
somewhat, the change is not enough to tolerate the latency introduced
by the path.

BUG=b:253295333
TEST=None

Change-Id: I71fdf4b309b3699b0a9805c27fa323ccf99a4e1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4395646
Tested-by: Per Åhgren <peah@chromium.org>
Reviewed-by: Ben Zhang <benzh@chromium.org>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>
Auto-Submit: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Per Åhgren <peah@chromium.org>