reven: floss: Add flex_bluetooth depency to Flex builds
We are seeing four Bluetooth adapters on ChromeOS Flex have issues with
the reading of vendor capabilities, and this is making it so that
Bluetooth is continually cycling off and on. The flex_bluetooth
package in platform2 was added to address this. It contains an
executable `process_flex_bluetooth_overrides` that will execute before
`bluetoothd` starts so that Floss is able to pick up the overrides.
More details about the original issue regarding reading of vendor
capabilities can be found at b/257423916.

BUG=b:273335825, b:273321276, b:273582512, b:273969466
TEST=On a laptop with one of the affected Bluetooth adapters,
     `emerge-reven chromeos-bsp-reven && cros deploy $DUT
     chromeos-bsp-reven`. Verify that the file
     `floss_reven_overrides.conf` exists at
     `/var/lib/bluetooth/sysprops.conf.d/` and contains
     ```
     [Sysprops]
     bluetooth.core.le.vendor_capabilities.enabled=false
     ```
     . Also verify that Bluetooth is able to be turned on successfully.

Cq-Depend: chromium:4436426
Change-Id: I4ba89b58d6b5471373650ddf8b1807a4a3373e8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4404447
Commit-Queue: Adrian Chan <adchan@google.com>
Tested-by: Adrian Chan <adchan@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>