trogdor/strongbad: boost ARCVM process by using boost-arcvm
Set boost-arcvm value to 0.82, which will result in a uclamp.min boost
between 58% for the fastest models (2.55Ghz), up to 70% for the slowest
models (2.1Ghz).

This should be enough to make the performance video test cases to pass,
which are especially sensitive to a latency issues.

BUG=b:217825939
TEST=GtsExoPlayerTestCases

Cq-Depend: chromium:4459615
Change-Id: I74baa41b18f20291eb99d7d51b84c4896ac03197
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4460231
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Marcin Wojtas <mwojtas@google.com>
Reviewed-by: Marcin Wojtas <mwojtas@google.com>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>