overlays: Enable Chrome debug fission
Chrome build fails in the latest uprev with the link errors in
.debug_info:
"relocation R_X86_64_32 out of range: 4295031334 is not in [0,
4294967295]"

This is not the first time when it happened and may likely happen on
other boards. The solution going forward is to use debug fission (or
split DWARF), see b/271459198.
We want to start from the minimal set of boards to unblock chrome uprev
since we haven't used debug fission on amd64 boards before and may
discover some issues.

Enable debug fission on volteer, eve, atlas.

BUG=b:279487992
TEST=cq

Cq-Depend: chromium:4472599
Change-Id: Icfe03176a035f319c0191eb32a9dc7712f7ca0b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4475296
Owners-Override: Yulun Wu <yulunwu@chromium.org>
Tested-by: Denis Nikitin <denik@chromium.org>
Reviewed-by: Yulun Wu <yulunwu@chromium.org>