brya: enable crosvm-swap
Set the crosvm-swap USE flag on brya to facilitate teamfooding. This
only compiles in support for the feature - it is enabled at runtime by
Chrome flags.

This feature will eventually be rolled out to all boards and enabled by
default, but for now it's being tested/supported on a board-by-board
basis.

BUG=b:268280946
TEST=CQ

Change-Id: Ia000b853687119d44354f14be22b40f1cc38a2d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4600589
Tested-by: David Stevens <stevensd@chromium.org>
Reviewed-by: YH Lin <yueherngl@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>
Reviewed-by: Shin Kawamura <kawasin@google.com>