amd64-host: Add bootstrap profile for bazel host SDK
As part of the bazel work, we are going to generate a new SDK that
contains the latest primordial host tools. We want this SDK to be as
slim as possible so we will be disabling USE flags to minimize the size
of the SDK. We also need to include the toolchains:base profile so we
can pull in the correct config so we can build the cross-* packages
using bazel.

Instead of using the existing `base` profile, I'm making a new one to
avoid the risk that we interfere with the existing SDK builder.

BUG=b:278728702
TEST=BOARD=amd64-generic bazel build @portage//internal/sdk/stage2/host

Change-Id: I40b092c809ecd718b01defbb6dadc64446e11525
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4603841
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>