Rex: Add intel_psr USE flag
This USE flag is to tell that this board supports Intel PSR(platform
service record) feature. With this flag, we can use it to implement a
Tast software dependency, so we can run Tast on supported boards. Or we
can use it to determine whether build the intel-psrtool or not.

BUG=b:276396727, b:287564676
TEST=`emerge-rex --info` shows intel_psr

Change-Id: I76e020647f59270f91a8fa330ee4525ba1f80f72
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4619790
Reviewed-by: YH Lin <yueherngl@chromium.org>
Commit-Queue: Kerker Yang <kerker@chromium.org>
Tested-by: Kerker Yang <kerker@chromium.org>