baseboard-brya: update sof-topology to migrate to cavs2.5-001
The production build is moved to sof-private cavs2.5-001-drop-stable
branch.

In addition, the strategy is applied for creating the symlink for
each ADL topology file to a RPL copy beforehand, due to the fact that
the RPL refresh is planned for many ADL-P devices.

BUG=b:289374146
TEST=`emerge-brya sof-topology` and check files installed under
     /build/brya/lib/firmware/intel/sof-tplg/

Cros-Add-Test-Suites: bvt-tast-cq-audio
Cros-Add-TS-Boards-BuildTarget: brya
Change-Id: I45bb126d7be4d976b08b4a118041f91de02a0679
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4664955
Reviewed-by: Terry Cheong <htcheong@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: Pin-chih Lin <johnylin@chromium.org>
Commit-Queue: Terry Cheong <htcheong@chromium.org>