brask: update sof-topology to migrate to cavs2.5-001
The production build is moved to sof-private cavs2.5-001-drop-stable
branch.

In addition, the strategy is applied for creating the symlink for
each ADL topology file to a RPL copy beforehand, due to the fact that
the RPL refresh is planned for many ADL-P devices.

BUG=b:289374146\
TEST=`emerge-brask sof-topology` and check files installed under
         /build/brask/lib/firmware/intel/sof-tplg/

Change-Id: Id9893f20d622860fe2a2c840a5a8c3c189b0240d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4664956
Tested-by: Pin-chih Lin <johnylin@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Terry Cheong <htcheong@chromium.org>
Auto-Submit: Pin-chih Lin <johnylin@chromium.org>
Commit-Queue: Pin-chih Lin <johnylin@chromium.org>