arm64-generic: Add vm-optimized profile
This profile builds an image that is bootable in QEMU and GCP Compute
Engine. It's useful for tests that require an arm64 CrOS userspace.

The presubmit check for board_specific_setup.sh has been skipped as the
script is not meant to be run directly. Instead it's called by another
script cros_make_image_bootable which sets up all the required variables
for it. The remaining checks are passing.

BUG=b:286335746
TEST=boot arm64-generic image with vm-optimized profile

Cq-Depend: chromium:4627386, chromium:4713910, chromium:4659470
Change-Id: Ie40690d1c350b7dcaeaec1db03eb313fa52476dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4627683
Tested-by: Yi Xie <yixie@chromium.org>
Commit-Queue: Yi Xie <yixie@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>