host prebuilt.conf: Update SDK versions for crossdev fix
At present, libcxx is failing to select the actual libcxx
version and instead is using the unstable 9999 version. We
have to reset the libcxx prebuilts, along with the actual
crossdev patch which introduced the bug.

This CL updates the SDK for those prebuilts.

These versions come from the chromiumos-sdk tryjob on branch
stabilize-15564.B.

Log run:
https://ci.chromium.org/ui/p/chromeos/builders/general/Full/b8773922233632834689/overview

BUG=b:293898274
TEST=CQ and tested from SDK stabilize branch

Cq-Depend: chromium:4749012
Change-Id: Ibd21271cb11f543de2f4a3638bcddffd114b420c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4749049
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: George Burgess <gbiv@chromium.org>