dedede: Remove linux-sources overlay
The dedede specific linux-sources overlay was added to allow custom
patches to be applied before they were ready for inclusion in
chromeos-kernel-5_4. See https://crrev.com/c/2080697. The patches were
eventually included and most of the overlay was removed in
https://crrev.com/c/2538264. However, the linux-sources override was
not.

Remove the dedede specific linux-sources ebuild as it is no longer
needed and is causing compilation issues when trying to build v6.1 of
the kernel.

BUG=b:302040955
TEST=-`KBUILD_BUILD_TIMESTAMP="Jan 5, 2021" emerge-dedede -v
      chromeos-kernel-5_4` with and without overlay removed.
     -`diff vmlinuz-with-overlay mvlinux-without-overlay` returns 0
     -Load and boot kernel

Change-Id: I74cce8f2c8bc5b0e2ac63ed6390aff479af7359a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4895838
Tested-by: Mark Hasemeyer <markhas@google.com>
Commit-Queue: Mark Hasemeyer <markhas@google.com>
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>