baseboard-brya: update sof-topology for merging brask package
Given that overlay-brask is based on baseboard-brya now, it is not legal
to define another sof-topology package in overlay-brask while there is
one already in baseboard-brya. It we really did it that way (we actually
did), if sof-topology ebuild version of brask is less than the one of
brya, the latter will be loaded by brask anyway. On the other hand, if
the version of brask is larger, the building process will report errors
on it.

A possible solution is to make sof-topology a virtual package, however I
didn't see a merit for sof-topology package which should prefer to have
the clear classification by each, rather than to have the chance to be
overwritten. On that account, this CL removes sof-topology package of
brask while merging to the one in baseboard-brya.

BUG=b:274567389
TEST=`emerge-brask sof-topology` and check files installed under
     /build/brask/lib/firmware/intel/sof-tplg/

Change-Id: I72ff9249a140f673ba46cea60fcdd9fc418932b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4902668
Tested-by: Pin-chih Lin <johnylin@chromium.org>
Commit-Queue: Pin-chih Lin <johnylin@chromium.org>
Reviewed-by: Terry Cheong <htcheong@chromium.org>
Auto-Submit: Pin-chih Lin <johnylin@chromium.org>