baseboard-dedede: add recovery KERN-C
I downloaded dedede recovery_image.bin of 15632.0.0, dump KERN-A. The
non-zero part of KERN-A is about 32MB. So shrinking KERN-A to 200 MB
should be safe.

Note that the order of partitions matters, `cros build-image` will check
the order, and p6 must go before p2.

BUG=b:289116981
TEST=CQ
TEST=setup_board --board=dedede
     cros build-packages --board=dedede
     cros build-image --board=dedede base
     ./mod_image_for_recovery.sh --board=dedede --image=...
     sudo losetup /dev/loop0 ...
     sudo diff /dev/loop0p2 /dev/loop0p6  # they are the same

Change-Id: I1b26956a42662eb1f4d12bb9942c34431aac0e18
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4904257
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Stimim Chen <stimim@chromium.org>